[DF-3834] Fixed crashes and memory leaks
Former-commit-id: 3e86953c46d9874b31c812ee1a3756fe8c1e3922

+++ Tools/ResourceEditor/Classes/Deprecated/ParticlesEditorNodeNameHelper.cpp
@@ -38,7 +38,7 @@ String ParticlesEditorNodeNameHelper::GetBaseName(const String& name)
     String numberName;
     // Get numbers string at the end of entity name
     const char* cName = baseName.c_str();
-    for (int i = baseName.length() - 1; i >= 0; --i)
+    for (int i = static_cast<int>(baseName.length()) - 1; i >= 0; --i)
     {
         char a = cName[i];
         if (a >= '0' && a <= '9')
